+++ src/components/app/app.test.js
@@ -9,7 +9,6 @@ const host = {
     `An independent House, strategically located between Rembrand Square and National Opera, but where the bustle of the city comes to rest in this alley flowery and colorful.`
   ],
 };
-
 const placeCards = [
   {
     link: `/`,
@@ -56,6 +55,7 @@ const placeCards = [
     ],
     isPremium: true,
     rating: 4,
+    locations: [52.3909553943508, 4.85309666406198],
     host,
   },
   {
@@ -103,6 +103,7 @@ const placeCards = [
     ],
     isPremium: true,
     rating: 4,
+    locations: [51.3909553943508, 4.85309666406198],
     host,
   },
   {
@@ -150,6 +151,7 @@ const placeCards = [
     ],
     isPremium: true,
     rating: 4,
+    locations: [53.3909553943508, 4.85309666406198],
     host,
   },
   {
@@ -197,15 +199,49 @@ const placeCards = [
     ],
     isPremium: true,
     rating: 4,
+    locations: [54.3909553943508, 4.85309666406198],
     host,
+  },
+];
+const reviews = [
+  {
+    text: `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.`,
+    userRate: 4,
+    userName: `Max`,
+    date: `December 24, 2018`,
+    id: `1`,
+  },
+  {
+    text: `The building is green and from 18th century.`,
+    userRate: 3,
+    userName: `Max`,
+    date: `July 11, 2018`,
+    id: `2`,
+  },
+  {
+    text: `The building is green and from 18th century. A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.`,
+    userRate: 2,
+    userName: `Max`,
+    date: `May 14, 2018`,
+    id: `3`,
+  },
+  {
+    text: `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam. The building is green and from 18th century.`,
+    userRate: 1,
+    userName: `Max`,
+    date: `April 24, 2019`,
+    id: `4`,
   },
 ];
 
 it(`<App /> snapshot`, () => {
   const tree = renderer
     .create(<App
+      reviews = {reviews}
       placeCards = {placeCards}
-    />)
+    />,
+    {createNodeMock: () => document.createElement(`section`)}
+    )
     .toJSON();
 
   expect(tree).toMatchSnapshot();
