+++ src/components/card-detail/сard-detailtest.js
@@ -1,4 +1,4 @@
-import Property from './property';
+import CardDetail from './cardDetail';
 
 const host = {
   avatar: `img/avatar-angelina.jpg`,
@@ -9,7 +9,36 @@ const host = {
     `An independent House, strategically located between Rembrand Square and National Opera, but where the bustle of the city comes to rest in this alley flowery and colorful.`
   ],
 };
-
+const reviews = [
+  {
+    text: `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.`,
+    userRate: 4,
+    userName: `Max`,
+    date: `December 24, 2018`,
+    id: `1`,
+  },
+  {
+    text: `The building is green and from 18th century.`,
+    userRate: 3,
+    userName: `Max`,
+    date: `July 11, 2018`,
+    id: `2`,
+  },
+  {
+    text: `The building is green and from 18th century. A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.`,
+    userRate: 2,
+    userName: `Max`,
+    date: `May 14, 2018`,
+    id: `3`,
+  },
+  {
+    text: `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam. The building is green and from 18th century.`,
+    userRate: 1,
+    userName: `Max`,
+    date: `April 24, 2019`,
+    id: `4`,
+  },
+];
 const placeData = {
   link: `/`,
   img: `img/apartment-01.jpg`,
@@ -60,8 +89,9 @@ const placeData = {
 
 it(`<Property /> snapshot`, () => {
   const tree = renderer
-    .create(<Property
+    .create(<CardDetail
       placeData = {placeData}
+      reviews = {reviews}
     />)
     .toJSON();
