+++ src/components/map/map.jsx
@@ -1,10 +1,6 @@
 import leaflet from 'leaflet';
 
-const icon = leaflet.icon({
-  iconUrl: `img/pin.svg`,
-  iconSize: [30, 30]
-});
-const zoom = 12;
+const ZOOM = 12;
 
 class Map extends PureComponent {
   constructor(props) {
@@ -13,15 +9,20 @@ class Map extends PureComponent {
   }
 
   componentDidMount() {
-    const {offerCords, city} = this.props;
+    const {locations, city} = this.props;
 
-    const map = leaflet.map(this.mapRef.current.id, {
+    const icon = leaflet.icon({
+      iconUrl: `img/pin.svg`,
+      iconSize: [30, 30]
+    });
+
+    const map = leaflet.map(this.mapRef.current, {
       center: city,
-      zoom,
+      zoom: ZOOM,
       zoomControl: false,
       marker: true
     });
-    map.setView(city, zoom);
+    map.setView(city, ZOOM);
 
     leaflet
       .tileLayer(`https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png`, {
@@ -29,7 +30,7 @@ class Map extends PureComponent {
       })
       .addTo(map);
 
-    offerCords.forEach((item) => {
+    locations.forEach((item) => {
       leaflet
         .marker(item, {icon})
         .addTo(map);
@@ -37,17 +38,20 @@ class Map extends PureComponent {
   }
 
   render() {
+    const {isDetailsPage} = this.props;
+
     return (<>
-      <section ref={this.mapRef} className="cities__map map" id="map" />
+      <section ref={this.mapRef} className={`${isDetailsPage ? `property` : `cities`}__map map`} id="map" />
     </>);
   }
 }
 
 Map.propTypes = {
-  offerCords: PropTypes.arrayOf(
+  locations: PropTypes.arrayOf(
       PropTypes.arrayOf(PropTypes.number.isRequired)
   ).isRequired,
   city: PropTypes.array.isRequired,
+  isDetailsPage: PropTypes.bool.isRequired,
 };
 
 export default Map;
