+++ src/components/map/map.test.js
@@ -1,21 +1,24 @@
-import Map from 'map';
+import Map from './map';
 
-const mockLocations = [
-  [52.3909553943508, 4.85309666406198],
-  [52.369553943508, 4.85309666406198],
-  [52.3909553943508, 4.929309666406198],
-  [52.3809553943508, 4.939309666406198]
-];
-const mockCityLocation = [52.38333, 4.9];
+describe(`<Map/>`, () => {
 
-it(`<Map /> snapshot`, () => {
-  const tree = renderer
-    .create(<Map
-      offerCords={mockLocations}
-      city={mockCityLocation}
-    />, {
-      createNodeMock: () => document.createElement(`div`)
-    }).toJSON();
+  const mockLocations = [
+    [52.3909553943508, 4.85309666406198],
+    [52.369553943508, 4.85309666406198],
+    [52.3909553943508, 4.929309666406198],
+    [52.3809553943508, 4.939309666406198]
+  ];
+  const mockCityLocation = [52.38333, 4.9];
 
-  expect(tree).toMatchSnapshot();
+  it(`should render`, () => {
+    const tree = renderer
+      .create(<Map
+        offerCords={mockLocations}
+        city={mockCityLocation}
+      />,
+      {createNodeMock: () => document.createElement(`section`)}
+      ).toJSON();
+
+    expect(tree).toMatchSnapshot();
+  });
 });
